add video slide support
